+++ source_gen/source_gen/source_gen/test/src/unformatted_code_generator.dart
@@ -13,17 +13,9 @@ class UnformattedCodeGenerator extends Generator {
 
   static const formattedCode = '''
 void hello() => print('hello');
-
-void x() {
-  <String>[].add('y');
-}
 ''';
 
   static const unformattedCode = '''
 void hello ()=>  print('hello');
-
-void x() {
-  <String>[]..add('y');
-}
 ''';
 }